requirements.txt
aio-pika==9.5.4
aiormq==6.8.1
asgiref==3.8.1
Django==5.1.4
django-cors-headers==4.6.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
exceptiongroup==1.2.2
idna==3.10
multidict==6.1.0
pamqp==3.3.0
pika==1.3.2
propcache==0.2.1
psycopg2==2.9.10
python-decouple==3.8
python-dotenv==1.0.1
sqlparse==0.5.3
yarl==1.18.3
